add jobs

+++ apps/paper_jobs/context.py
@@ -7,7 +7,7 @@ def args():
     parser.add_argument('-e', '--epochs', type=int, help='epochs count')
     parser.add_argument('-b', '--batch', type=int, help='batch count')
     parser.add_argument('-r', '--round', type=int, help='number of rounds')
-    parser.add_argument('-cr', '--clients_ratio', type=float, help='selected client percentage for fl')
+    parser.add_argument('-cr', '--clients_ratio', type=int, help='selected client percentage for fl')
     parser.add_argument('-lr', '--learn_rate', type=float, help='learn rate')
     parser.add_argument('-t', '--tag', type=str, help='tag to save the results')
     return parser.parse_args()
